Petits bugs dans LargeFile
git-svn-id: http://caml.inria.fr/svn/ocaml/trunk@4475 f963ae5c-01c2-4b8c-9fe0-0dff7051ff02

+++ otherlibs/threads/unix.ml
@@ -259,7 +259,7 @@ external link : string -> string -> unit = "unix_link"
 
 module LargeFile =
   struct
-    external lseek : file_descr -> int64 -> seek_command -> int = "unix_lseek_64"
+    external lseek : file_descr -> int64 -> seek_command -> int64 = "unix_lseek_64"
     external truncate : string -> int64 -> unit = "unix_truncate_64"
     external ftruncate : file_descr -> int64 -> unit = "unix_ftruncate_64"
     type stats =
